traceroute overhaul for sanity
